Trabalho Concluido

+++ SPA/js/api.js
@@ -5,6 +5,12 @@ class Api {
         return data;
     }
 
+    async getUser(id) {
+        const response = await fetch(`https://localhost:5001/api/user/${id}`);
+        const data = response.json();       
+        return data;
+    }    
+
     async  postUser(data) {
         const response = await fetch("https://localhost:5001/api/auth/register", {
             method: 'POST',
@@ -12,18 +18,36 @@ class Api {
                 'Content-type': 'application/json'
             },
             body: JSON.stringify(data)
-        })
+        });
+        const responseData = await response.json();
+        return responseData;
+    }
+
+    async putUser(id, data) {
+        const response = await fetch(`https://localhost:5001/api/user/${id}`, {
+            method: 'PUT',
+            headers: {
+                'Content-type':'application/json'
+            },
+            body: JSON.stringify(data)
+        });
         const responseData = await response.json();
         return responseData;
     }
 
     async deleteUser(id) {
-        const response =  await fetch (`https://localhost:5001/api/user/${id}`,{
-            method:'DELETE',
-            headers: {
-                'Content-type':'application/json'
-            }
-        });
-        window.location.reload();
+        if(confirm('Tem certeza que deseja excluír?'))
+        {
+            await fetch (`https://localhost:5001/api/user/${id}`,{
+                method:'DELETE',
+                headers: {
+                    'Content-type':'application/json'
+                }
+            });
+            window.location.reload();
+        }
+        else{
+            return false;
+        }
     }
 }
